scalars docs
