+++ package.json
@@ -14,7 +14,8 @@
     "format": "ts-node scripts/local.ts format",
     "format:fix": "ts-node scripts/local.ts  format --fix=true",
     "code-gen": "ts-node scripts/local.ts code-gen",
-    "lint": "curl -sSL https://raw.github.com/ndmitchell/hlint/master/misc/run.sh | sh -s ."
+    "lint": "curl -sSL https://raw.github.com/ndmitchell/hlint/master/misc/run.sh | sh -s .",
+    "test:server": "stack test --fast  morpheus-graphql-server  --ta"
   },
   "devDependencies": {
     "@types/glob": "^8.0.0",
